+++ src/Orckestra.Composer.Website/App_Data/Razor/Composer/SingleCheckout/Steps/ReviewCart.cshtml
@@ -100,7 +100,7 @@
                 @RemoveCartItemMobileBtn()
             </div>
             <div v-if="item.RecurringOrderProgramName" class="col-11  offset-2 offset-sm-1  order-8">
-                @RecurringOrderHelpers.CartItemRecurringOrderProgram()
+                @RecurringOrderHelpers.CartItemRecurringOrderProgram("SingleCheckout.ChangeRecurringFrequency")
             </div>
         </div>
     </div>
